% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlore.R
\name{db_list_}
\alias{db_list_}
\title{Create db_list_xyz functions.}
\usage{
db_list_(obj)
}
\arguments{
\item{obj}{String used with \code{switch} in the function
\code{\link{db_list_sql_}} to select appropriate SQL query. Choices include
\code{"schemas"}, \code{"tables"}, \code{"views"}, and \code{"tvs"}.}
}
\value{
A function that can be assigned a custom variable name.
}
\description{
\code{db_list_} is a function factory to create functions that returns
properties of a Microsoft SQL Server database.
}
\details{
\code{db_list_} is used along with \code{\link{db_list_sql_}} to create
functions.
}
\section{Warning}{
 The intent is not to use this function directly.  Instead
  use one of the functions
  \itemize{
    \item \code{\link{db_list_schemas}}
    \item \code{\link{db_list_tables}}
    \item \code{\link{db_list_views}}
    \item \code{\link{db_list_tvs}}
  }
}
\seealso{
\href{http://adv-r.had.co.nz/Functional-programming.html#closures}{Closures}
}

