% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlore.R
\name{db_list_schemas}
\alias{db_list_schemas}
\title{List schemas.}
\usage{
db_list_schemas(db_con, db_name, db_schema)
}
\arguments{
\item{db_con}{Database server connection object.}

\item{db_name}{Database name.}

\item{db_schema}{Database schema.}
}
\value{
A character vector of all schemas.
}
\description{
List all schemas in a Microsoft SQL Server database.
}
\examples{
\dontrun{
con <- RSQLServer::src_sqlserver(server = "dbserver_dbname", file = "~/.sql.yaml")
dbserver_schemas <- db_list_schemas(db_con = con)
}

}

