% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlore.R
\name{db_list_tables}
\alias{db_list_tables}
\title{List tables.}
\usage{
db_list_tables(db_con, db_name, db_schema)
}
\arguments{
\item{db_con}{Database server connection object.}

\item{db_name}{Database name.}

\item{db_schema}{Database schema.}
}
\value{
A character vector of all tables.
}
\description{
List all tables in a Microsoft SQL Server database.
}
\section{Warning}{

Will not return views.  If tables and views are desired, use
\code{\link{db_list_tvs}}.
}
\examples{
\dontrun{
con <- RSQLServer::src_sqlserver(server = "dbserver_dbname", file = "~/.sql.yaml")
dbserver_tables <- db_list_tables(db_con = con, db_name = "dbname")
}
}

