% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlore.R
\name{db_list_tvs}
\alias{db_list_tvs}
\title{List tables and views.}
\usage{
db_list_tvs(db_con, db_name, db_schema)
}
\arguments{
\item{db_con}{Database server connection object.}

\item{db_name}{Database name.}

\item{db_schema}{Database schema.}
}
\value{
A character vector of all tables and views.
}
\description{
List all tables and views in a Microsoft SQL Server database.
}
\examples{
\dontrun{
con <- RSQLServer::src_sqlserver(server = "dbserver_dbname", file = "~/.sql.yaml")
dbserver_tvs <- db_list_tvs(db_con = con,
                            db_name = "dbname",
                            db_schema = "dbschema")
}
}

