% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlore.R
\name{db_list_views}
\alias{db_list_views}
\title{List views.}
\usage{
db_list_views(db_con, db_name, db_schema)
}
\arguments{
\item{db_con}{Database server connection object.}

\item{db_name}{Database name.}

\item{db_schema}{Database schema.}
}
\value{
A character vector of all views.
}
\description{
List all views in a Microsoft SQL Server database.
}
\section{Warning}{

Will not return tables.  If tables and views are desired, use
\code{\link{db_list_tvs}}.
}
\examples{
\dontrun{
con <- RSQLServer::src_sqlserver(server = "dbserver_dbname", file = "~/.sql.yaml")
dbserver_views <- db_list_views(db_con = con,
                                db_name = "dbname",
                                db_schema = "dbschema")
}
}

