% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{df_diff}
\alias{df_diff}
\title{Test how dataframes are the same / different.}
\usage{
df_diff(x, y)
}
\arguments{
\item{x}{First dataframe.}

\item{y}{Second dataframe.}
}
\value{
Test results and differences are printed to standard out.
}
\description{
\code{df_diff} prints out how dataframes are the same and how they are
different.
}
\details{
Tests include
\itemize{
  \item{Equal # of rows}
  \item{Equal # of columns}
  \item{Row names are the same}
  \item{Column names are the same}
  \item{Class(es) is / are the same}
  \item{Column types are the same - work in progress}
  \item{Values within each column are the same}
}
\code{df_diff} is unstable at the moment and should be considered a work in
progress.
}
\examples{
x <- tibble::frame_data(
 ~col1, ~col2,
 1, "a",
 2, "b",
 3, "c"
)

y <- data.frame(col1 = 1:3, col2 = c("a", "b", "d"), col3 = c(10, 11, 12))

df_diff(x, y)

}

