% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{exact_cols}
\alias{exact_cols}
\title{List exact names of columns.}
\usage{
exact_cols(tbl, cols)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} or dataframe.}

\item{cols}{A character vector of column names.}
}
\value{
A character vector of column names with the appropriate case.
}
\description{
\code{exact_cols} returns the exact column names, including the case.
}
\details{
This particularly useful when working with Microsoft SQL Server tables as
the column names may be case sensitive.  This is also useful when working
with table metadata and the exact column names are required for matching
purposes.

\link[dplyr]{tbl_vars} does not return case sensitive columns, thus the
the need for this function.
}
\examples{
## create tibble
my_tibble <- tibble::tribble(
  ~MixedCase, ~lowercase, ~UPPERCASE, ~aLtCaSe,
  1, 11, 111, 1111,
  2, 22, 222, 2222,
  3, 33, 333, 3333
)

## column names to search for, ignoring case
my_cols <- c("mixedcase", "altcase")

my_cols_case <- my_tibble \%>\%
  exact_cols(my_cols)

## particularly useful when combined with dplyr::one_of
my_tibble \%>\%
  dplyr::select(dplyr::one_of(my_cols_case))

}

