% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{get_xml_text}
\alias{get_xml_text}
\title{Get XML Text}
\usage{
get_xml_text(x, xpath)
}
\arguments{
\item{x}{XML document: a literal XML document, a URL, or a string.}

\item{xpath}{A string containing a xpath (1.0) expression.}
}
\value{
XML text as a string or \code{NA_character_} if an error occurs.
}
\description{
Return an XML text value if provided the document and Xpath.
}
\details{
Serves as a convenience function.  Utilizes \code{\link{safe_xml_find_first}}
and \code{\link{safe_xml_text}} and will return \code{NA_character_} if an
error occurs.
}
\examples{
xml_doc <- "
<root_tag>
  <another_tag>
    <var_text>10</var_text>
    <var_attr value=\\"10\\"></var_attr>
  </another_tag>
</root_tag>
"

ten <- get_xml_text(xml_doc,
                    "//root_tag/another_tag/var_text")
identical(ten, "10")

}

