% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{is_linux}
\alias{is_linux}
\title{Is Linux.}
\usage{
is_linux()
}
\value{
A logical vector indicating whether the OS is Linux.
}
\description{
Is the OS Linux.
}
\seealso{
Other is_os: \code{\link{is_osx}}, \code{\link{is_windows}}
}

