% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{is_os_}
\alias{is_os_}
\title{Generate OS check.}
\usage{
is_os_(os_string)
}
\arguments{
\item{os_string}{String containing the OS to check for.
Choices include
\itemize{
  \item \code{"os x"}
  \item \code{"linux"}
  \item \code{"windows"}
}}
}
\value{
A function that checks the type of OS.
}
\description{
\code{is_os_} returns a function to check the operating system in use.
}
\details{
The intent is not to use this function directly.  Instead use one of the
functions
\itemize{
  \item \code{\link{is_linux}}
  \item \code{\link{is_osx}}
  \item \code{\link{is_windows}}
}
}
\examples{

}
\seealso{
\href{http://adv-r.had.co.nz/Functional-programming.html#closures}{Closures}
}

