% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{is_osx}
\alias{is_osx}
\title{Is OSX.}
\usage{
is_osx()
}
\value{
A logical vector indicating whether the OS is OSX.
}
\description{
Is the OS OSX.
}
\seealso{
Other is_os: \code{\link{is_linux}},
  \code{\link{is_windows}}
}

