% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{is_windows}
\alias{is_windows}
\title{Is Windows}
\usage{
is_windows()
}
\value{
A logical vector indicating whether the OS is Windows.
}
\description{
Is the OS Windows.
}
\seealso{
Other is_os: \code{\link{is_linux}}, \code{\link{is_osx}}
}

