% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{match_cols}
\alias{match_cols}
\title{Find a column name.}
\usage{
match_cols(tbl, patt, ...)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} or dataframe.}

\item{patt}{Regular expression pattern to pass to \code{grepl}.}

\item{...}{Parameters to pass along to \code{grepl}.}
}
\value{
A character vector of column names matching the regex.
}
\description{
\code{match_cols} returns matching column names from a table.  The matching
is performed using regular expressions and \link[base]{grepl}.
}
\details{
Ignores case by default.  Can be overridden by passing \code{ignore.case} via
\code{...}.
}
\examples{
## dataframe to convert
key <- c(1000L, 2000L, 3000L, 4000L, 1000L)
amount <- c("46.41", "118.11", "84.68", "493.59", "51.10")

test_df <- data.frame(key, amount)

# column names
my_col <- "^am"

# use
test_df \%>\%
  match_cols(my_col)

}

