% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{pkgs_df}
\alias{pkgs_df}
\title{Packages Dataframe.}
\usage{
pkgs_df()
}
\value{
Returns a \href{https://cran.rstudio.com/package=tibble}{tibble}
containing a subset of information about the packages installed.
}
\description{
\code{pkgs_df} returns a \link[tibble]{tibble} containing a subset of
information about the packages installed.
}
\examples{
pkgs <- pkgs_df()
pkgs

}
\seealso{
\link{check_install}
}

