% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%pmatches\%}
\alias{\%pmatches\%}
\title{Perl Matches.}
\usage{
lhs \%pmatches\% rhs
}
\arguments{
\item{lhs}{The character vector to search; the haystack.}

\item{rhs}{The \code{perl} regular expression; the needle.}
}
\value{
Logical vector indicating a match.
}
\description{
Infix operator to check if a character vector matches a \code{perl} regular
expression.
}
\examples{
virginica_df <- iris \%>\%
  dplyr::filter(Species \%pmatches\% "^vir")

}

