% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{prior_install}
\alias{prior_install}
\title{Install prior version of a package.}
\usage{
prior_install(pkg, pkg_version)
}
\arguments{
\item{pkg}{Character vector of the package to install.}

\item{pkg_version}{Character vector of the package version to install.}
}
\value{
There is no return as the function is called for its side effects.
}
\description{
\code{prior_install} will install a prior version of a package from source.
\code{prior_install} will first check that the version supplied exists by
querying for the tar.gz file from the
\href{https://cran.rstudio.com/src/contrib/Archive}{CRAN archive}.  If the
\code{tar.gz} file exists, it will attempt to install it.
}
\examples{
\dontrun{
prior_install(pkg = "tidytext", pkg_version = "0.1.0")
}

}
\seealso{
\itemize{
  \item \link{check_install}
  \item \href{https://rstudio.github.io/packrat/}{Packrat}
  \item \href{https://github.com/robertzk/lockbox}{Lockbox}
}
}

