% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{safe_xml_attr}
\alias{safe_xml_attr}
\title{Safe \code{xml2} extract attr function}
\usage{
safe_xml_attr(...)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{ns}{Optionally, a named vector giving prefix-url pairs, as produced
by \code{\link[xml2]{xml_ns}}. If provided, all names will be explicitly
qualified with the ns prefix, i.e. if the element \code{bar} is defined
in namespace \code{foo}, it will be called \code{foo:bar}. (And
similarly for atttributes). Default namespaces must be given an explicit
name.}

\item{attr}{Name of attribute to extract.}

\item{default}{Default value to use when attribute is not present.}
}
\value{
A character vector or \code{NA_character_} if there is an error.
}
\description{
Safe version of \code{xml2::xml_attr} function.
}
\details{
The safe version is helpful when iterating over a set of XML documents
and the schemas are inconsistent.  Rather than throwing errors, the function
will return \code{NA_character_} when unable to extract from a node.

Suggested resources for XPath are
\itemize{
  \item \url{https://developer.mozilla.org/en-US/docs/Web/XPath/Functions}
  \item \url{http://www.w3schools.com/xsl/xpath_syntax.asp}
}
}
\examples{
xml_doc <- "
<root_tag>
  <another_tag>
    <var_text>10</var_text>
    <var_attr value=\\"10\\"></var_attr>
  </another_tag>
</root_tag>
"

var_attr <- safe_xml_attr(
  safe_xml_find_first(xml2::read_xml(xml_doc),
                      xpath = "//root_tag/another_tag/var_attr")$result,
  "value")

# returns "10" without error
var_attr$result
is.null(var_attr$error)

var_attr2 <- safe_xml_attr(
  safe_xml_find_first(xml2::read_xml(xml_doc),
                      xpath = "//root_tag/another_tag/var_attr")$result,
  "value2")

# does not return a value but returns an error
is.null(var_attr2$result)
var_attr2$error

}
\seealso{
\itemize{
  \item \link[xml2]{xml_attr}
  \item \link[xml2]{xml_text}
}
}

