% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{safe_xml_find}
\alias{safe_xml_find}
\alias{safe_xml_find_all}
\alias{safe_xml_find_one}
\title{Safe \code{xml2} find functions}
\usage{
safe_xml_find_one(...)

safe_xml_find_all(...)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{xpath}{A string containing a xpath (1.0) expression.}

\item{ns}{Optionally, a named vector giving prefix-url pairs, as produced
by \code{\link[xml2]{xml_ns}}. If provided, all names will be explicitly
qualified with the ns prefix, i.e. if the element \code{bar} is defined
in namespace \code{foo}, it will be called \code{foo:bar}. (And
similarly for atttributes). Default namespaces must be given an explicit
name.}
}
\value{
\code{safe_xml_find_all} and \code{safe_xml_find_one} will either
return a nodeset (or node) or return \code{NA_character_} if there is not
a match.
}
\description{
Safe version of \code{xml2} find functions
}
\details{
The safe versions are helpful when iterating over a set of XML documents
and the schemas are inconsistent.  Rather than throwing errors, the safe find
function will return \code{NA_character_} when unable to find a node.

Suggested resources for XPath are
\itemize{
  \item \url{https://developer.mozilla.org/en-US/docs/Web/XPath/Functions}
  \item \url{http://www.w3schools.com/xsl/xpath_syntax.asp}
}
}
\examples{
var_text_node <- safe_xml_find_one(xml2::read_xml(xml_doc),
                                   xpath = "//root_tag/another_tag/var_text")

# returns a nodeset without an error
var_text_node$result
is.null(var_text_node$error)

var_text2_node <- safe_xml_find_one(xml2::read_xml(xml_doc),
                                    xpath = "//root_tag/another_tag/var_text2")

# does not return a nodeset but returns an error
is.null(var_text2_node$result)
var_text2_node$error

}
\seealso{
\itemize{
  \item \link[xml2]{xml_find_all}
}
}

