% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{safe_xml_text}
\alias{safe_xml_text}
\title{Safe \code{xml2} extract text function}
\usage{
safe_xml_text(...)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{trim}{If \code{TRUE} will trim leading and trailing spaces.}
}
\value{
A character vector or \code{NA_character_} if there is an error.
}
\description{
Safe version of \code{xml2::xml_text} function.
}
\details{
The safe version is helpful when iterating over a set of XML documents
and the schemas are inconsistent.  Rather than throwing errors, the function
will return \code{NA_character_} when unable to extract from a node.

Suggested resources for XPath are
\itemize{
  \item \url{https://developer.mozilla.org/en-US/docs/Web/XPath/Functions}
  \item \url{http://www.w3schools.com/xsl/xpath_syntax.asp}
}
}
\examples{
xml_doc <- "
<root_tag>
  <another_tag>
    <var_text>10</var_text>
    <var_attr value=\\"10\\"></var_attr>
  </another_tag>
</root_tag>
"

var_text <- safe_xml_text(
  safe_xml_find_one(xml2::read_xml(xml_doc),
                    xpath = "//root_tag/another_tag/var_text")$result)

# returns "10" without error
var_text$result
is.null(var_text$error)

var_text2 <- safe_xml_text(
  safe_xml_find_one(xml2::read_xml(xml_doc),
                    xpath = "//root_tag/another_tag/var_text2")$result)

# does not return a value but returns an error
is.null(var_text2$result)
var_text2$error

}
\seealso{
\itemize{
  \item \link[xml2]{xml_attr}
  \item \link[xml2]{xml_text}
}
}

