% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-types.R
\name{spec_to_df}
\alias{spec_to_df}
\title{Create a dataframe from a \link[readr]{cols} specification.}
\usage{
spec_to_df(spec)
}
\arguments{
\item{spec}{A \link[readr]{cols} specification to be written as a dataframe.}
}
\value{
A \link[tibble]{tibble} representation of a \link[readr]{cols}
specification.
}
\description{
Given a \link[readr]{cols} specificaton, convert it to a dataframe.
}
\details{
One use case is to write the returned dataframe as a \code{csv} in lieu of
writing out a \link[readr]{cols} specification as an \code{rds} object.  This
is useful for cases when reading a \link[readr]{cols} specification as an
\code{rds} throws an error.

The inverse operation is \link{spec_from_df}.
}

