% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlore.R
\name{sql_col_names}
\alias{sql_col_names}
\title{Get exact names of columns from SQL Server.}
\usage{
sql_col_names(cols, tbl)
}
\arguments{
\item{cols}{character vector of column names}

\item{tbl}{table object (tibble)}
}
\value{
A character vector of column names with the appropriate case.
}
\description{
Because Microsoft SQL Server column names can be case sensitive, get the exact column names
as SQL Server stores them.
}
\examples{
\dontrun{
# connection and tibble objects
con <- RSQLServer::src_sqlserver(server = "dbserver_dbname", file = "~/.sql.yaml")
my_tbl <- dplyr::tbl(con, sql("SELECT * FROM [dbname].[dbschema].[mytable]"))

# get column names
my_cols <- sql_col_names(c("mixedCasecol", "anothermixedCol"))
# use
my_tbl \%>\% select(my_cols)
}

}

