% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlore.R
\name{sql_list_}
\alias{sql_list_}
\title{Create sql_list_xyz functions.}
\usage{
sql_list_(obj)
}
\arguments{
\item{obj}{String used with \code{switch} in the function
\code{\link{sql_list_sql_}} to select appropriate SQL query. Choices include
\itemize{
  \item \code{"schemas"}
  \item \code{"tables"}
  \item \code{"views"}
  \item \code{"tvs"}
}}
}
\value{
A function that can be assigned a custom variable name.
}
\description{
\code{sql_list_} is a function factory to create functions that return
properties of a Microsoft SQL Server database.
}
\details{
\code{sql_list_} is used along with \code{\link{sql_list_sql_}} to create
functions.
}
\section{Warning}{
 The intent is not to use this function directly.  Instead
  use one of the functions
  \itemize{
    \item \code{\link{sql_list_schemas}}
    \item \code{\link{sql_list_tbls}}
    \item \code{\link{sql_list_views}}
    \item \code{\link{sql_list_tvs}}
  }
}
\seealso{
\href{http://adv-r.had.co.nz/Functional-programming.html#closures}{Closures}
}

