% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-explore.R
\name{sql_list_cols}
\alias{sql_list_cols}
\title{Get exact names of columns from SQL Server.}
\usage{
sql_list_cols(tbl, cols)
}
\arguments{
\item{tbl}{table object (tibble)}

\item{cols}{character vector of column names}
}
\value{
A character vector of column names with the appropriate case.
}
\description{
\code{sql_list_cols} returns Microsoft SQL Server column names with the
appropriate case.  Because column names can be case sensitive, get the exact
column names as SQL Server stores them.
}
\examples{
\dontrun{
# connection and tibble objects
con <- RSQLServer::src_sqlserver(server = "dbserver_dbname", file = "~/.sql.yaml")
my_tibble <- dplyr::tbl(con, dplyr::sql("SELECT * FROM [dbname].[dbschema].[dbtable]"))

# column names
my_cols <- c("mixedCasecol", "anothermixedCASECol")

# use
my_tibble \%>\%
  sql_list_cols(my_cols)
}

}
\seealso{
Other sql_lists: \code{\link{sql_list_schemas}},
  \code{\link{sql_list_tbls}}, \code{\link{sql_list_tvs}},
  \code{\link{sql_list_views}}
}

