% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-explore.R
\name{sql_list_schemas}
\alias{sql_list_schemas}
\title{List schemas.}
\usage{
sql_list_schemas(db_con, db_name)
}
\arguments{
\item{db_con}{Database server connection object.}

\item{db_name}{Database name.}
}
\value{
A character vector of all schemas.
}
\description{
List all schemas in a Microsoft SQL Server database.
}
\examples{
\dontrun{
con <- RSQLServer::src_sqlserver(server = "dbserver_dbname", file = "~/.sql.yaml")
dbserver_schemas <- sql_list_schemas(db_con = con,
                                     db_name = "dbname")
}

}
\seealso{
Other sql_lists: \code{\link{sql_list_cols}},
  \code{\link{sql_list_tbls}}, \code{\link{sql_list_tvs}},
  \code{\link{sql_list_views}}
}

