% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlore.R
\name{sql_list_sql_}
\alias{sql_list_sql_}
\title{Generate SQL query.}
\usage{
sql_list_sql_(db_con, db_name, db_schema, obj)
}
\arguments{
\item{db_con}{Database server connection object.}

\item{db_name}{Database name.}

\item{db_schema}{Database schema.}

\item{obj}{String used with \code{switch} to select appropriate SQL query.
Choices include
\itemize{
  \item \code{"schemas"}
  \item \code{"tables"}
  \item \code{"views"}
  \item \code{"tvs"}
}}
}
\value{
An SQL query as a string.
}
\description{
\code{sql_list_sql_} returns an SQL query as a string based on the parameters
passed to it.
}
\details{
\code{sql_list_sql_} is used along with the function factory
\code{\link{sql_list_}} to create functions that return properties of a
Microsoft SQL Server database.
}
\section{Warning}{
 The intent is not to use this function directly.  Instead
  use one of the functions
  \itemize{
    \item \code{\link{sql_list_schemas}}
    \item \code{\link{sql_list_tbls}}
    \item \code{\link{sql_list_views}}
    \item \code{\link{sql_list_tvs}}
  }
}

