% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-explore.R
\name{sql_list_tbls}
\alias{sql_list_tbls}
\title{List tables.}
\usage{
sql_list_tbls(db_con, db_name, db_schema)
}
\arguments{
\item{db_con}{Database server connection object.}

\item{db_name}{Database name.}

\item{db_schema}{Database schema.}
}
\value{
A character vector of all tables.
}
\description{
\code{sql_list_tbls} lists all tables in a Microsoft SQL Server database.
}
\details{
Will not return views.  If tables and views are desired, use
\code{\link{sql_list_tvs}}.
}
\examples{
\dontrun{
con <- RSQLServer::src_sqlserver(server = "dbserver_dbname", file = "~/.sql.yaml")
dbserver_tables <- sql_list_tbls(db_con = con,
                                 db_name = "dbname",
                                 db_schema = "dbschema")
}

}
\seealso{
Other sql_lists: \code{\link{sql_list_schemas}},
  \code{\link{sql_list_tvs}}, \code{\link{sql_list_views}}
}

