% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlore.R
\name{sql_list_tvs}
\alias{sql_list_tvs}
\title{List tables and views.}
\usage{
sql_list_tvs(db_con, db_name, db_schema)
}
\arguments{
\item{db_con}{Database server connection object.}

\item{db_name}{Database name.}

\item{db_schema}{Database schema.}
}
\value{
A character vector of all tables and views.
}
\description{
\code{sql_list_tvs} lists all tables and views in a Microsoft SQL Server
database.
}
\examples{
\dontrun{
con <- RSQLServer::src_sqlserver(server = "dbserver_dbname", file = "~/.sql.yaml")
dbserver_tvs <- sql_list_tvs(db_con = con,
                             db_name = "dbname",
                             db_schema = "dbschema")
}

}
\seealso{
Other sql_lists: \code{\link{sql_list_cols}},
  \code{\link{sql_list_schemas}},
  \code{\link{sql_list_tbls}}, \code{\link{sql_list_views}}
}

