% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{types_df}
\alias{types_df}
\title{Column names and types in a dataframe.}
\usage{
types_df(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A \link[tibble]{tibble} with the column names and types.
}
\description{
Given a dataframe, \code{types_df} returns a \link[tibble]{tibble} with the
column names and types.
}
\examples{
(mtcars \%>\%
   types_df())

}

