% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml_get_attr}
\alias{xml_get_attr}
\title{Get XML Attribute.}
\usage{
xml_get_attr(x, xpath, extract_value)
}
\arguments{
\item{x}{XML document: a literal XML document, a URL, or a string}

\item{xpath}{A string containing a xpath (1.0) expression.}

\item{extract_value}{A string containing the attribute name to extract.}
}
\value{
XML node attribute as a string or \code{NA_character_} if an error
occurs.
}
\description{
Return an XML node attribute if provided the document, XPath, and attribute
name.
}
\details{
Serves as a convenience function.  Utilizes \link[xml2]{xml_find_first}
and \link[xml2]{xml_attr} and will return \code{NA_character_} if the XML
attribute cannot be located or an error occurs.

Because the function uses \link[xml2]{xml_find_first}, only the first match
is returned.
}
\examples{
xml_doc <- "
<root_tag>
  <another_tag>
    <var_text>10</var_text>
    <var_attr value=\\"10\\"></var_attr>
  </another_tag>
</root_tag>
"

ten <- xml_get_attr(xml_doc,
                    "//root_tag/another_tag/var_attr",
                    "value")
identical(ten, "10")

}

