% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-install.R
\name{check_install}
\alias{check_install}
\title{Check and install packages.}
\usage{
check_install(pkgs)
}
\arguments{
\item{pkgs}{List or vector of packages to check and optionally install.}
}
\value{
Invisibly returns the input pkgs list or vector.  The function is
called for its side effects.
}
\description{
\code{check_install} checks if a package is already installed.  If it is not
installed, then install the package (from CRAN).
}
\examples{
\dontrun{
req_pkgs <- list(
  "dplyr",
  "purrr",
  "rprojroot"
)
check_install(req_pkgs)
}

}

