% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{get_xml_text}
\alias{get_xml_text}
\title{Get XML Text.}
\usage{
get_xml_text(x, xpath)
}
\arguments{
\item{x}{XML document: a literal XML document, a URL, or a string.}

\item{xpath}{A string containing a xpath (1.0) expression.}
}
\value{
XML text as a string or \code{NA_character_} if an error occurs.
}
\description{
Return an XML text value if provided the document and Xpath.
}
\details{
Serves as a convenience function.  Utilizes \link[xml2]{xml_find_first}
and \link[xml2]{xml_text} and will return \code{NA_character_} if the XML
text cannot be located or an error occurs.

Because the function uses \link[xml2]{xml_find_first}, only the first match
is returned.
}
\examples{
xml_doc <- "
<root_tag>
  <another_tag>
    <var_text>10</var_text>
    <var_attr value=\\"10\\"></var_attr>
  </another_tag>
</root_tag>
"

ten <- get_xml_text(xml_doc,
                    "//root_tag/another_tag/var_text")
identical(ten, "10")

}

