% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-types.R
\name{spec_to_df}
\alias{spec_to_df}
\title{Convert \link[readr]{spec} to a dataframe}
\usage{
spec_to_df(spec)
}
\arguments{
\item{spec}{A \link[readr]{spec} to be written as a dataframe}
}
\value{
A \link[tibble]{tibble} representation of a \link[readr]{spec}.
}
\description{
Take a \link[readr]{spec} and converts it to a dataframe.  The dataframe can
then be written out as a CSV in lieu of writing out an actual \code{rds}
object.  This is useful for cases when reading a \link[readr]{spec} that was
saved as an \code{rds} throws an error.
}
\details{
After reading the specification in CSV format back into \code{R}, it can be
converted into an actual specification using \link{col_spec_update}.  See the
example for methodology.
}

